+++ db/likes.ts
@@ -1,6 +1,8 @@
 import db from "./db.ts";
+import { anyUnescaped } from "./dbMethods.ts";
 
 export async function queryLikes(tweet_id: string) {
+  if (anyUnescaped(tweet_id)) return false;
   return await db(async (client) =>
     (await client.queryArray`
         SELECT author_id
@@ -11,6 +13,7 @@ export async function queryLikes(tweet_id: string) {
 }
 
 export async function queryLikedBy(tweet_id: string, user_id: string) {
+  if (anyUnescaped(tweet_id, user_id)) return false;
   return await db(async (client) =>
     (await client.queryArray`
         SELECT author_id
@@ -21,6 +24,7 @@ export async function queryLikedBy(tweet_id: string, user_id: string) {
 }
 
 export async function createLike(tweet_id: string, user_id: string) {
+  if (anyUnescaped(tweet_id, user_id)) return false;
   await db(async (client) =>
     await client.queryArray`
             INSERT INTO likes (tweet_id, author_id)
@@ -31,6 +35,7 @@ export async function createLike(tweet_id: string, user_id: string) {
 }
 
 export async function deleteLike(tweet_id: string, user_id: string) {
+  if (anyUnescaped(tweet_id, user_id)) return false;
   await db(async (client) =>
     await client.queryArray`
                 DELETE FROM likes
